Quick changes
Moved files into folders. Introduced a .gitignore that should cover every possible extraneous file. Updated readme format.
